requirements.txt
facebook-sdk==3.1.0
openai==1.3.0
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
